Fix Spring Boot 1.4 upgrade
Re-add `spring-boot-starter` dependency to the `movies-data-access`
components. In the previous version the `spring-boot-starter` dependency
was used in `testCompile`. During the upgrade this was changed to
`spring-boot-starter-test` meaning that the snakeyaml library was no
longer included. This has the knock on effect of meaning the
`application.yml` test config was ignored.

Update `RepositoryTestConfiguration` to directly import
`JdbcTemplateAutoConfiguration`. In Spring Boot 1.3 the `JdbcTemplate`
was defined in `DataSourceAutoConfiguration` but in Spring Boot 1.4
this changed. See Spring Boot issue 6449 for details.